Organize src files into subdirectories
