+++ docs/misc/claude_desktop_config_example.json
@@ -2,7 +2,7 @@
   "mcpServers": {
     "dockashell": {
       "command": "node",
-      "args": ["/Users/anzax/Projects/dockashell/src/mcp-server.js"],
+      "args": ["/Users/anzax/Projects/dockashell/src/mcp/mcp-server.js"],
       "env": {
         "NODE_ENV": "production"
       }
