+++ src/core/security.js
